Stash and re-apply tests completed.

+++ Tests/Editor/Core/Build/Test_StashGitChangesTests.cs
@@ -10,8 +10,8 @@ namespace ExtenityTests.Build
 {
     public class Test_StashGitChangesTests
     {
-        private static readonly string _dummyRepoPath = Path.Combine(Application.persistentDataPath,"DummyRepo").FixDirectorySeparatorChars();
-        private static readonly string _dummySubmodulePath = Path.Combine(Application.persistentDataPath,"Submodule").FixDirectorySeparatorChars();
+        private static readonly string _dummyRepoPath = Path.Combine(Application.persistentDataPath, "DummyRepo").FixDirectorySeparatorChars();
+        private static readonly string _dummySubmodulePath = Path.Combine(Application.persistentDataPath, "Submodule").FixDirectorySeparatorChars();
 
         private GitCommandRunner _commandRunner;
         [OneTimeSetUp]
@@ -19,27 +19,28 @@ namespace ExtenityTests.Build
         {
             CreateDummyRepo(_dummyRepoPath);
             CreateDummyRepo(_dummySubmodulePath);
-            
+
             CreateDummyFiles(_dummyRepoPath);
             CreateDummyFiles(_dummySubmodulePath);
-            
+
             CommitDummyFiles(_dummyRepoPath);
             CommitDummyFiles(_dummySubmodulePath);
             
-           
             CreateDummyFiles(_dummySubmodulePath);
-            
-            AddSubmoduleToTheDummyRepo(_dummyRepoPath,_dummySubmodulePath);
-            
+
+            AddSubmoduleToTheDummyRepo(_dummyRepoPath, _dummySubmodulePath);
+
             CommitDummyFiles(_dummyRepoPath);
+
+            CreateDummyFiles(_dummyRepoPath);
             
-            CreateDummyFiles(Path.Combine(_dummyRepoPath,"Sub"));
+            CreateDummyFiles(Path.Combine(_dummyRepoPath, "Sub"));
         }
 
         [OneTimeTearDown]
         public void TearDown()
         {
-            //DeleteDummyRepoFolder();
+            DeleteDummyRepoFolders();
         }
 
         private void CreateDummyRepo(string path)
@@ -50,9 +51,9 @@ namespace ExtenityTests.Build
             }
             _commandRunner = new GitCommandRunner(path);
 
-           var output = _commandRunner.Run($"init . -b master", out int exitCode);
-           
-           Debug.Log(output);
+            var output = _commandRunner.Run($"init . -b master", out int exitCode);
+
+            Debug.Log(output);
         }
 
         private void CreateDummyFiles(string repoPath)
@@ -60,45 +61,63 @@ namespace ExtenityTests.Build
             var directoryInfo = new DirectoryInfo(repoPath);
 
             var files = directoryInfo.GetFiles();
-            
+
             for (int i = files.Length; i < files.Length + 5; i++)
             {
-                File.WriteAllText(Path.Combine(repoPath,$"File{i}.txt"), i.ToString());
+                File.WriteAllText(Path.Combine(repoPath, $"File{i}.txt"), i.ToString());
             }
         }
         private void CommitDummyFiles(string repoPath)
         {
             _commandRunner = new GitCommandRunner(repoPath);
             _commandRunner.Run("add .", out int exitCode);
-            _commandRunner.Run("commit -m 'test'", out  exitCode);
-            
-            Assert.AreEqual(0,exitCode);
+            _commandRunner.Run("commit -m 'test'", out exitCode);
+
+            Assert.AreEqual(0, exitCode);
         }
 
         private void AddSubmoduleToTheDummyRepo(string repo, string submodule)
         {
             _commandRunner = new GitCommandRunner(repo);
-            _commandRunner.Run($"submodule add {submodule}  Sub",out var exitCode);
-            _commandRunner.Run("submodule update --init --force --remote",out  exitCode);
-            Assert.AreEqual(0,exitCode);
+            _commandRunner.Run($"submodule add {submodule}  Sub", out var exitCode);
+            _commandRunner.Run("submodule update --init --force --remote", out exitCode);
+            Assert.AreEqual(0, exitCode);
 
         }
 
-        [Test]
+        [Test, Order(0)]
         public void StashAllGitLocalChanges()
         {
-            BuildTools.StashAllLocalGitChanges(_dummyRepoPath,null,false);
+            BuildTools.StashAllLocalGitChanges(_dummyRepoPath, null, false);
         }
-        [Test]
+        [Test, Order(1)]
         public void StashAllGitLocalChangesIncludingSubmodules()
         {
             BuildTools.StashAllLocalGitChanges(_dummyRepoPath);
         }
-
-        private void DeleteDummyRepoFolder()
+        [Test, Order(2)]
+        public void ApplyLastStash()
         {
-            Directory.Delete(_dummyRepoPath,true);
-            Directory.Delete(_dummySubmodulePath,true);
+            BuildTools.ApplyLastGitStash(_dummyRepoPath,false);
+        }
+        [Test, Order(3)]
+        public void ApplyLastStashIncludingSubmodules()
+        {
+            StashAllGitLocalChangesIncludingSubmodules();
+            BuildTools.ApplyLastGitStash(_dummyRepoPath);
+        }
+
+
+        private void DeleteDummyRepoFolders()
+        {
+            if (Directory.Exists(_dummySubmodulePath))
+            {
+                Directory.Delete(_dummySubmodulePath,true);
+            }
+            if (Directory.Exists(_dummyRepoPath))
+            {
+                Directory.Delete(_dummyRepoPath,true);
+            }
         }
     }
 }
